LayoutTests/fast/exclusions/shape-outside-floats/shape-outside-floats-ellipse-margin-right.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="../resources/subpixel-utils.js"></script>
<script>
    if (window.internals)
        window.internals.settings.setCSSExclusionsEnabled(true);
</script>

<style>
#container-border {
    border: 1px solid black; 
    margin: 10px; 
    width: 500px;
}

#container {
    font: 20px/1 Ahem, sans-serif;
    width: 500px;
    height: 300px;
    text-align: right;
  }

#float-right {
    float: right;
    position: relative;
    overflow: visible;
    -webkit-shape-outside: ellipse(200px, 100px, 200px, 100px);
    -webkit-shape-margin: 25px;
}

#svg-shape-ellipse {
    position: absolute;
    display:block;
    top: 0px;
    left: -50px;
    width: 500px;
    height: 300px;
}
</style>

<body>
<p>The right edges of the three black squares should follow the outer ellipse boundary and each square should appear on a subsequent line.</p>
<div id="container-border">
    <div id="container">
        <div id="float-right">
            <svg id="svg-shape-ellipse" xmlns="http://www.w3.org/2000/">
                <ellipse cx="250" cy="100" rx="200" ry="100" fill="green"></ellipse>
                <ellipse cx="250" cy="100" rx="225" ry="125" fill="none" stroke="green"></ellipse>
            </svg>
        </div>
        <span id="s1">X</span></br><span id="s2">X</span><br/><span id="s3">X</span>
    </div>
</div>
</body>
<script>
function elementRect(elementId) 
{
    var s = document.getElementById("container").getBoundingClientRect();
    var r = document.getElementById(elementId).getBoundingClientRect();
    return {right: (r.left - s.left) + r.width, top: r.top - s.top, width: r.width, height: r.height};
}

// You'll find the equation for the X intercept of an elliptical arc here (among other places):
// http://hansmuller-webkit.blogspot.com/2012/07/computing-horizonal-rounded-rectangle.html
// 
// Note also: only the "ellipse(200px, 100px, 200px, 100px)" itself fits within the container element.  The margin
// ellipse extends above and to the left by another 25 pixels.  The +25 adjustment to Y accounts for this.


function marginEllipseRightXIntercept(y, cx, rx, ry)
{
    var containerWidth = document.getElementById("container").getBoundingClientRect().width;
    return String( SubPixelLayout.roundLineRight(containerWidth - (cx + rx * Math.sqrt(1 - Math.pow((ry - (y + 25)) / ry, 2)))) );
}

// In the ".right" tests below, it's the lower edge of the 20px Ahem glyph that that defines the element rect's left edge. 
// That's why the Y coordinate in each of those tests is 20 greater than the ".top" test.

SubPixelLayout.initSubPixelLayout();

shouldBe("elementRect('s1').top", "0");
shouldBe("elementRect('s1').right", marginEllipseRightXIntercept(20, 200, 225, 125));

shouldBe("elementRect('s2').top", "20");
shouldBe("elementRect('s2').right", marginEllipseRightXIntercept(40, 200, 225, 125));

shouldBe("elementRect('s3').top", "40");
shouldBe("elementRect('s3').right", marginEllipseRightXIntercept(60, 200, 225, 125));

</script>
</html>
